feat(drone): 完善无人机系统功能和扩展机制
- 在编译阶段过滤编辑器专用节点，只保留可执行节点
- 添加高级导航和EU接口无人机升级配方
- 将OpenComputers依赖改为可选编译依赖
- 重构无人机动作注册事件为最终类并简化API
- 新增无人机货物满载策略枚举类型
- 添加扩展不可用诊断代码
- 重构无人机编辑器覆盖层实现，支持分组、注释和时间线
- 为无人机端点添加存储容量和资源检查方法
- 实现无人机端点网络的清理和移除功能
- 添加物流相关的能量消耗计算方法
- 扩展无人机实体服务攻击方法参数
- 增强无人机执行器注册表的安全性检查
- 实现无人机扩展可用性检查和显示元数据
- 创建无人机扩展显示和注册机制
- 添加无人机扩展注册中心统一管理所有扩展组件

+++ src/main/java/com/drppp/drtech/compat/opencomputers/OpenComputersPage.java
@@ -11,7 +11,7 @@ public final class OpenComputersPage {
         if (values == null || values.isEmpty()) return Collections.emptyList();
         int safePage = Math.max(0, page);
         int safeSize = Math.max(1, Math.min(64, pageSize));
-        int from = Math.min(values.size(), safePage * safeSize);
+        int from = (int) Math.min((long) values.size(), (long) safePage * safeSize);
         int to = Math.min(values.size(), from + safeSize);
         return Collections.unmodifiableList(new ArrayList<>(values.subList(from, to)));
     }
